backend/requirements.txt
asgiref
attrs
autopep8
django-cors-headers
Deprecated
Django
django-extensions
django-environ
iniconfig
numpy
packaging
pip
pluggy
py
pycodestyle
pydantic
pymongo==3.12.3
pyparsing
pytest
pytest-django
python-dotenv
pytz
PyYAML
setuptools
sqlparse
toml
tomli
typing_extensions
wheel
wrapt
djangorestframework
mysqlclient
colorlog
